lnx1-tst2

+++ variadic_functions/0-sum_them_all.c
@@ -19,7 +19,7 @@ int sum_them_all(const unsigned int n, ...)
 		return (0);
 	}
 
-	for (int i = 0; i < n; i++)
+	for (int i = 0; i < (int) n; i++)
 	{
 		topla += va_arg(args, int);
 	}
